solve build warnings.

+++ nvtc/TrCountingGraph.cpp
@@ -31,7 +31,7 @@ struct GET_LENGTH_ARGS {
 struct BATCH_R4_ARGS {
     TrCountingGraph* G;
     const char* file_name;
-    int length;
+    int64_t length;
     int from;
     int step;
 };
@@ -336,7 +336,7 @@ void* loadbatch_R4(void* args) {
     struct BATCH_R4_ARGS* batch_r4_args = (struct BATCH_R4_ARGS*) args;
     TrCountingGraph* G = batch_r4_args->G;
     const char* file_name = batch_r4_args->file_name;
-    int length = batch_r4_args->length;
+    int64_t length = batch_r4_args->length;
     int from = batch_r4_args->from;
     int step = batch_r4_args->step;
     FILE* pFile = fopen(file_name, "rb");
